refactor: 프로젝트 폴더 구조 변경
원래 목적이었던 기능별로 확실하게 역할이 구별되도록 다시 분리
