+++ Matcheap/app/src/main/java/com/sandy/seoul_matcheap/adapters/RegionSpinnerAdapter.kt
@@ -1,8 +1,7 @@
-package com.sandy.seoul_matcheap.ui.home
+package com.sandy.seoul_matcheap.adapters
 
 import com.sandy.seoul_matcheap.R
 import com.sandy.seoul_matcheap.databinding.ItemRvSpinnerBinding
-import com.sandy.seoul_matcheap.ui.common.BaseListAdapter
 import com.sandy.seoul_matcheap.util.constants.TYPE_CATEGORY
 import com.sandy.seoul_matcheap.util.constants.TYPE_REGION
