+++ Matcheap/app/src/main/java/com/sandy/seoul_matcheap/adapters/SearchHistoryListAdapter.kt
@@ -1,9 +1,8 @@
-package com.sandy.seoul_matcheap.ui.search
+package com.sandy.seoul_matcheap.adapters
 
 import com.sandy.seoul_matcheap.R
 import com.sandy.seoul_matcheap.data.store.entity.SearchHistory
 import com.sandy.seoul_matcheap.databinding.ItemRvHistoryBinding
-import com.sandy.seoul_matcheap.ui.common.BaseListAdapter
 
 /**
  * @author SANDY
