+++ Matcheap/app/src/main/java/com/sandy/seoul_matcheap/adapters/StoreCountPagerAdapter.kt
@@ -1,9 +1,8 @@
-package com.sandy.seoul_matcheap.ui.more
+package com.sandy.seoul_matcheap.adapters
 
 import com.sandy.seoul_matcheap.R
 import com.sandy.seoul_matcheap.data.store.dao.StoreTotalCount
 import com.sandy.seoul_matcheap.databinding.ItemPagerStoreCountBinding
-import com.sandy.seoul_matcheap.ui.common.BaseListAdapter
 
 /**
  * @author SANDY
