+++ Matcheap/app/src/main/java/com/sandy/seoul_matcheap/adapters/StoreMenuAdapter.kt
@@ -1,9 +1,9 @@
-package com.sandy.seoul_matcheap.ui.store
+package com.sandy.seoul_matcheap.adapters
 
 import com.sandy.seoul_matcheap.R
 import com.sandy.seoul_matcheap.data.store.entity.Menu
 import com.sandy.seoul_matcheap.databinding.ItemRvMenuBinding
-import com.sandy.seoul_matcheap.ui.common.BaseListAdapter
+import com.sandy.seoul_matcheap.adapters.BaseListAdapter
 
 /**
  * @author SANDY
