+++ Matcheap/app/src/main/java/com/sandy/seoul_matcheap/ui/MainActivity.kt
@@ -12,9 +12,8 @@ import androidx.navigation.findNavController
 import androidx.navigation.ui.setupWithNavController
 import com.sandy.seoul_matcheap.R
 import com.sandy.seoul_matcheap.databinding.ActivityMainBinding
-import com.sandy.seoul_matcheap.ui.common.SoftKeyboardDetectorView
+import com.sandy.seoul_matcheap.util.custom.SoftKeyboardDetectorView
 import com.sandy.seoul_matcheap.ui.map.MapFilterFragment
-import com.sandy.seoul_matcheap.util.*
 import com.sandy.seoul_matcheap.util.constants.*
 import dagger.hilt.android.AndroidEntryPoint
