+++ Matcheap/app/src/main/java/com/sandy/seoul_matcheap/ui/more/notice/NoticeFragment.kt
@@ -6,9 +6,10 @@ import androidx.paging.LoadState
 import androidx.paging.PagingData
 import com.sandy.seoul_matcheap.MatcheapApplication.Companion.showToastMessage
 import com.sandy.seoul_matcheap.R
+import com.sandy.seoul_matcheap.adapters.NoticeTitleAdapter
 import com.sandy.seoul_matcheap.data.matcheap.Notice
 import com.sandy.seoul_matcheap.databinding.FragmentNoticeBinding
-import com.sandy.seoul_matcheap.ui.common.BaseFragment
+import com.sandy.seoul_matcheap.ui.BaseFragment
 import com.sandy.seoul_matcheap.util.constants.MESSAGE_NETWORK_ERROR
 
 class NoticeFragment : BaseFragment<FragmentNoticeBinding>(R.layout.fragment_notice) {
