+++ Matcheap/app/src/main/java/com/sandy/seoul_matcheap/ui/store/StoreDetailsActivity.kt
@@ -22,10 +22,11 @@ import com.google.firebase.ktx.Firebase
 import com.naver.maps.geometry.LatLng
 import com.naver.maps.map.*
 import com.sandy.seoul_matcheap.R
+import com.sandy.seoul_matcheap.adapters.StoreMenuAdapter
 import com.sandy.seoul_matcheap.data.store.dao.StoreMapItem
 import com.sandy.seoul_matcheap.data.store.entity.StoreInfo
 import com.sandy.seoul_matcheap.databinding.ActivityStoreDetailsBinding
-import com.sandy.seoul_matcheap.ui.common.TouchFrameLayout
+import com.sandy.seoul_matcheap.util.custom.TouchFrameLayout
 import com.sandy.seoul_matcheap.ui.more.bookmark.BookmarkViewModel
 import com.sandy.seoul_matcheap.util.constants.*
 import com.sandy.seoul_matcheap.util.helper.MapUtils
