+++ Matcheap/app/src/main/java/com/sandy/seoul_matcheap/util/custom/SoftKeyboardDetectorView.java
@@ -1,4 +1,4 @@
-package com.sandy.seoul_matcheap.ui.common;
+package com.sandy.seoul_matcheap.util.custom;
 
 import android.app.Activity;
 import android.content.Context;
