+++ Matcheap/app/src/main/java/com/sandy/seoul_matcheap/util/module/AppModules.kt
@@ -26,8 +26,8 @@ import com.sandy.seoul_matcheap.data.store.repository.MapRepository
 import com.sandy.seoul_matcheap.data.store.repository.SearchRepository
 import com.sandy.seoul_matcheap.data.store.repository.StoreRepository
 import com.sandy.seoul_matcheap.util.helper.MapUtils
-import com.sandy.seoul_matcheap.ui.more.settings.notification.NotificationScheduler
-import com.sandy.seoul_matcheap.ui.more.settings.notification.NotificationSchedulerImpl
+import com.sandy.seoul_matcheap.notification.NotificationScheduler
+import com.sandy.seoul_matcheap.notification.NotificationSchedulerImpl
 import com.sandy.seoul_matcheap.util.*
 import com.sandy.seoul_matcheap.util.constants.*
 import dagger.Module
